Refactor Category model: move to a dedicated file and remove old implementation
