+++ backend/voting_platform/api/admin.py
@@ -1,4 +1,4 @@
 from django.contrib import admin
-from .models import Category
+from .models.category import Category
 
 admin.site.register(Category)
