+++ backend/voting_platform/api/models/category.py
