+++ backend/voting_platform/api/views.py
@@ -1,5 +1,5 @@
 from rest_framework import generics
-from .models import Category
+from .models.category import Category
 from .serializers import CategorySerializer
